Changes to get WindowsBase to build
