+++ src/Microsoft.DotNet.Wpf/src/WindowsBase/MS/Internal/IO/Packaging/CompoundFile/RightsManagementEncryptionTransform.cs
@@ -1093,7 +1093,11 @@ namespace MS.Internal.IO.Packaging.CompoundFile
             ContentUser user
             )
         {
+#if NET
             return string.Create(null, stackalloc char[128], $"{user.AuthenticationType}:{user.Name}");
+#else
+            return $"{user.AuthenticationType}:{user.Name}";
+#endif
         }
 
         /// <summary>
