+++ src/Microsoft.DotNet.Wpf/src/WindowsBase/WindowsBase.csproj
@@ -64,6 +64,7 @@
     <Compile Include="$(WpfSharedDir)\MS\Win32\UnsafeNativeMethodsTextServices.cs" />
     <Compile Include="$(WpfSharedDir)\MS\Win32\SafeNativeMethodsCLR.cs" />
     <Compile Include="$(WpfSharedDir)\MS\Win32\SafeNativeMethodsOther.cs" />
+    <Compile Include="$(WpfSharedDir)\System\ReferenceEqualityComparer.cs" />
     <Compile Include="$(WpfSharedDir)\System\Windows\Markup\ReflectionHelper.cs" />
     <Compile Include="$(WpfSharedDir)\System\Windows\Interop\OSVersionHelper.cs" />
     <Compile Include="$(WpfSharedDir)\System\Windows\Interop\OperatingSystemVersion.cs" />
@@ -422,11 +423,13 @@
 
   <ItemGroup>
     <PackageReference Include="System.Diagnostics.EventLog" Version="$(SystemDiagnosticsEventLogPackageVersion)" />
-    <PackageReference Include="System.Windows.Extensions" Version="$(SystemWindowsExtensionsPackageVersion)" />
+    <PackageReference Condition="'$(TargetFrameworkIdentifier)'!='.NETFramework'" Include="System.Windows.Extensions" Version="$(SystemWindowsExtensionsPackageVersion)" />
     <PackageReference Include="System.Security.Cryptography.Xml" Version="$(SystemSecurityCryptographyXmlPackageVersion)" />
     <PackageReference Include="System.Configuration.ConfigurationManager" Version="$(SystemConfigurationConfigurationManagerPackageVersion)" />
-    <PackageReference Include="$(SystemIOPackagingPackage)" Version="$(SystemIOPackagingVersion)" />
+    <PackageReference Condition="'$(TargetFrameworkIdentifier)'!='.NETFramework'" Include="$(SystemIOPackagingPackage)" Version="$(SystemIOPackagingVersion)" />
+    <PackageReference Condition="'$(TargetFrameworkIdentifier)'=='.NETFramework'" Include="$(SystemIOPackagingPackage)" Version="$(SystemIOPackagingVersion)" ExcludeAssets="compile" />
     <PackageReference Include="System.Security.Permissions" Version="$(SystemSecurityPermissionsPackageVersion)" />
+    <PackageReference Condition="'$(TargetFrameworkIdentifier)'=='.NETFramework'" Include="System.Memory" Version="$(SystemMemoryPackageVersion)" />
   </ItemGroup>
 
   <ItemGroup>
@@ -434,6 +437,12 @@
     <ProjectReference Condition="'$(TargetFrameworkIdentifier)'!='.NETFramework'" Include="$(WpfSourceDir)WindowsBase\ref\WindowsBase-ref.csproj">
       <ReferenceOutputAssembly>false</ReferenceOutputAssembly>
     </ProjectReference>
+  </ItemGroup>
+
+  <ItemGroup Condition="'$(TargetFrameworkIdentifier)'=='.NETFramework'">
+    <Reference Include="Accessibility" />
+    <Reference Include="System.IO.Compression" />
+    <Reference Include="$(PkgSystem_IO_Packaging)\lib\netstandard2.0\System.IO.Packaging.dll" />
   </ItemGroup>
 
   <!-- Including the Microsoft.TextTemplating.targets requires us to manually import Sdk.Targets because
@@ -445,4 +454,10 @@
   <Import Project="$(WpfCodeGenDir)AvTrace\GenAvMessages.targets" />
   <!-- DevDiv Bug 888747: Re-enable AvTrace codegen -->
   <!-- <Import Project="$(WpfCodeGenDir)DesignTimeTextTemplating.targets" /> -->
+
+  <Target Condition="'$(TargetFrameworkIdentifier)'=='.NETFramework'" Name="RemoveNetFxReferences" BeforeTargets="ResolveAssemblyReferences">
+    <ItemGroup>
+      <Reference Remove="WindowsBase" />
+    </ItemGroup>
+  </Target>
 </Project>
